+++ nitrite/src/test/java/org/dizitart/no2/collection/DocumentTest.java
@@ -208,12 +208,15 @@ public class DocumentTest {
     @Test
     public void getFields() {
         Set<String> fields = doc.getFields();
-        assertEquals(fields.size(), 5);
+        assertEquals(fields.size(), 8);
         assertTrue(fields.contains("location.address.line1"));
         assertTrue(fields.contains("location.address.line2"));
+        assertTrue(fields.contains("location.address.house"));
         assertTrue(fields.contains("location.city"));
         assertTrue(fields.contains("location.state"));
         assertTrue(fields.contains("score"));
+        assertTrue(fields.contains("objArray"));
+        assertTrue(fields.contains("category"));
     }
 
     @Test